feat(draft): log execute time info of each runner node
The feature works with batch mode and request, for online mode only. The
log appears in two parts:

1. After each runner executed, prints summary log. This include execute
   time and execute results
2. After the whole SQL executed, prints summary log of all runners. As a
   single long log, with the format `[Query Trace] summary of SQL:
   {sql_str}\nRunner plan:\n{plan_str}\n{runner summary}...`

The logs only available when option `enable_debug` is true. For testing
purpose, the option is turned on by default.

You may also see too much queries to 'GLOBAL_VARIABLES', in that case,
higher the option `--sync_deploy_stats_timeout`, to 10000000(10000
seconds).